src/test/java/seedu/recipe/storage/JsonSerializableRecipeBookTest.java
package seedu.recipe.storage;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertNotNull;
import static org.junit.jupiter.api.Assertions.assertTrue;
import static org.junit.jupiter.api.Assertions.fail;
import static seedu.recipe.testutil.Assert.assertThrows;

import java.nio.file.Path;
import java.nio.file.Paths;
import java.util.List;

import org.junit.jupiter.api.Test;

import seedu.recipe.commons.exceptions.IllegalValueException;
import seedu.recipe.commons.util.JsonUtil;
import seedu.recipe.model.ReadOnlyRecipeBook;
import seedu.recipe.model.RecipeBook;
import seedu.recipe.model.recipe.Recipe;
import seedu.recipe.testutil.TypicalRecipes;

public class JsonSerializableRecipeBookTest {

    private static final Path TEST_DATA_FOLDER = Paths.get("src", "test", "data", "JsonSerializableRecipeBookTest");
    private static final Path TYPICAL_RECIPE_FILE = TEST_DATA_FOLDER.resolve("typicalRecipeRecipeBook.json");
    private static final Path INVALID_RECIPE_FILE = TEST_DATA_FOLDER.resolve("invalidRecipeRecipeBook.json");
    private static final Path DUPLICATE_RECIPE_FILE = TEST_DATA_FOLDER.resolve("duplicateRecipeRecipeBook.json");

    @Test
    public void toModelType_typicalRecipesFile_success() throws Exception {
        JsonSerializableRecipeBook dataFromFile = JsonUtil.readJsonFile(TYPICAL_RECIPE_FILE,
                JsonSerializableRecipeBook.class).get();
        RecipeBook recipeBookFromFile = dataFromFile.toModelType();
        RecipeBook typicalRecipesRecipeBook = TypicalRecipes.getTypicalRecipeBook();
        assertEquals(recipeBookFromFile.toString(), typicalRecipesRecipeBook.toString());
    }

    /**
     * Tests if an invalid data file with improperly formatted arguments causes the appropriate control flow
     * and raises an IllegalArgumentException, from attempting to instantiate with invalid arguments.
     * @throws Exception
     */
    @Test
    public void toModelType_invalidRecipeFile_throwsIllegalValueException() throws Exception {
        JsonSerializableRecipeBook dataFromFile = JsonUtil.readJsonFile(INVALID_RECIPE_FILE,
                JsonSerializableRecipeBook.class).get();
        assertThrows(IllegalValueException.class, dataFromFile::toModelType);
    }

    @Test
    public void toModelType_duplicateRecipes_throwsIllegalValueException() throws Exception {
        JsonSerializableRecipeBook dataFromFile = JsonUtil.readJsonFile(DUPLICATE_RECIPE_FILE,
                JsonSerializableRecipeBook.class).get();
        assertThrows(IllegalValueException.class, JsonSerializableRecipeBook.MESSAGE_DUPLICATE_PERSON,
                dataFromFile::toModelType);
    }

    @Test
    public void test_constructor() {
        ReadOnlyRecipeBook r = TypicalRecipes.getTypicalRecipeBook();
        List<Recipe> rList = r.getRecipeList();
        JsonSerializableRecipeBook test = new JsonSerializableRecipeBook(r);
        assertNotNull(test);

        try {
            RecipeBook testBook = test.toModelType();
            List<Recipe> recipeList = testBook.getRecipeList();
            assertEquals(recipeList.size(), rList.size());
            for (int i = 0; i < recipeList.size(); i++) {
                assertTrue(recipeList.get(i).isSameRecipe(rList.get(i)));
            }
        } catch (IllegalValueException e) {
            fail();
        }
    }
}
